pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spreadbot"
version = "0.1.0"
description = "Maker-biased market making / spread capture bot for Bybit linear perpetuals"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "OpenAI Codex"}]
dependencies = [
  "ccxt>=4.4.80",
  "numpy>=1.26.0",
  "optuna>=4.3.0",
  "pandas>=2.2.0",
  "pyarrow>=20.0.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.8.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.2",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "src",
]
testpaths = [
  "tests",
]
